Move cache fns to from registry to archive
