+++ x-pack/plugins/ingest_manager/server/services/epm/archive/cache.ts
@@ -3,7 +3,7 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-import { pkgToPkgKey } from './index';
+import { pkgToPkgKey } from '../registry/index';
 
 const cache: Map<string, Buffer> = new Map();
 export const cacheGet = (key: string) => cache.get(key);
